Tagging improvements
Change TagID from smallint to int.  As tags are added/removed there is a risk of overflowing smallint.
GUI improvements - using a grid to display the tag names/values to select instead of checkedlistbox.
Update to system tagging for logging agent version separate.

+++ DBADashDB/dbo/Tables/InstanceTags.sql
@@ -1,13 +1,9 @@
 ﻿CREATE TABLE [dbo].[InstanceTags] (
     [Instance] [sysname] NOT NULL,
-    [TagID]    SMALLINT  NOT NULL,
+    [TagID]    INT  NOT NULL,
     CONSTRAINT [PK_InstanceTags] PRIMARY KEY CLUSTERED ([Instance] ASC, [TagID] ASC),
     CONSTRAINT [FK_InstanceTags_Tags] FOREIGN KEY ([TagID]) REFERENCES [dbo].[Tags] ([TagID])
 );
-
-
-
-
 GO
 CREATE UNIQUE NONCLUSTERED INDEX [InstanceTags_TagID]
     ON [dbo].[InstanceTags]([TagID] ASC, [Instance] ASC);
